backend/requirements.txt
# Django Core
Django==5.0.4
djangorestframework==3.15.1
django-environ==0.11.2
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1

# Database / PostGIS
psycopg2-binary==2.9.9
django-postgis==0.0.2

# Celery
celery[redis]==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
redis==5.0.3

# Geospatial
# Note: GDAL must be installed at the OS level (included in PostGIS install)
# On Ubuntu: sudo apt-get install gdal-bin libgdal-dev
shapely==2.0.4
polyline==2.0.2

# Route Optimization
ortools==9.9.3963

# Utilities
Pillow==10.3.0
python-dateutil==2.9.0
